Refactor `JwtAuthenticationFilter` package, improve dependency injection, update controller response types to `Map<String, Object>`, replace `ArrayList` with `Collections.emptyList`, and adjust `SecurityConfig` for updated filter location.
